Move implementations of the classes ExampleWiseStratification and LabelWiseStratification into cpp files.
